+++ api/plan-dft-c2r-1d.c
@@ -20,7 +20,7 @@
 
 #include "api.h"
 
-X(plan) X(plan_dft_c2r_1d)(unsigned int n, C *in, R *out, unsigned int flags)
+X(plan) X(plan_dft_c2r_1d)(unsigned long n, C *in, R *out, unsigned int flags)
 {
      return X(plan_dft_c2r)(1, &n, in, out, flags);
 }
