+++ api/plan-dft-r2c-1d.c
@@ -20,7 +20,7 @@
 
 #include "api.h"
 
-X(plan) X(plan_dft_r2c_1d)(unsigned int n, R *in, C *out, unsigned int flags)
+X(plan) X(plan_dft_r2c_1d)(unsigned long n, R *in, C *out, unsigned int flags)
 {
      return X(plan_dft_r2c)(1, &n, in, out, flags);
 }
